desktop-app/requirements.txt
opencv-python>=4.10.0
Pillow>=10.4.0
qrcode>=7.4.2
